tests/wpt/css-tests/css-flexbox-1_dev/html/flexbox_computedstyle_flex-shorthand-number.htm

<!DOCTYPE html>
<html><head><title>flexbox | computed style | flex: number</title>
<link href="http://opera.com" rel="author" title="Opera Software">
<link href="http://www.w3.org/TR/css-flexbox-1/#propdef-flex" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags">
<style>
body {
	color: red;
	flex: 666;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-grow"),
	          "666");
    assert_equals(getComputedStyle(body).getPropertyValue("flex-shrink"),
	          "1");
    assert_equals(getComputedStyle(body).getPropertyValue("flex-basis"),
	          "0px");
});
</script>
</body></html>
